requirements.txt
streamlit==1.38.0
pillow==10.4.0
pdf2image==1.17.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy==2.0.2
langdetect==1.0.9
python-dateutil==2.9.0.post0
requests==2.32.3
google-auth==2.34.0
pdfminer.six==20221105
